collapse unnecessary HarrisLaplace class into HarrisLaplaceFeatureDetector
